Overwritten by auto-generated version. Build test has passed.

+++ mplapack/reference/Rgbrfs.cpp
@@ -47,7 +47,7 @@ void Rgbrfs(const char *trans, INTEGER const n, INTEGER const kl, INTEGER const 
     REAL safe1 = 0.0;
     REAL safe2 = 0.0;
     INTEGER count = 0;
-    const REAL three = 3.0e+0;
+    const REAL three = 3.0;
     REAL lstres = 0.0;
     const REAL one = 1.0;
     INTEGER i = 0;
@@ -55,7 +55,7 @@ void Rgbrfs(const char *trans, INTEGER const n, INTEGER const kl, INTEGER const 
     INTEGER kk = 0;
     REAL xk = 0.0;
     REAL s = 0.0;
-    const REAL two = 2.0e+0;
+    const REAL two = 2.0;
     const INTEGER itmax = 5;
     INTEGER kase = 0;
     INTEGER isave[3];
@@ -164,9 +164,9 @@ void Rgbrfs(const char *trans, INTEGER const n, INTEGER const kl, INTEGER const 
         s = zero;
         for (i = 1; i <= n; i = i + 1) {
             if (work[i - 1] > safe2) {
-                s = max(s, REAL(abs(work[(n + i) - 1]) / work[i - 1]));
+                s = max(s, abs(work[(n + i) - 1]) / work[i - 1]);
             } else {
-                s = max(s, REAL((abs(work[(n + i) - 1]) + safe1) / (work[i - 1] + safe1)));
+                s = max(s, (abs(work[(n + i) - 1]) + safe1) / (work[i - 1] + safe1));
             }
         }
         berr[j - 1] = s;
@@ -246,7 +246,7 @@ void Rgbrfs(const char *trans, INTEGER const n, INTEGER const kl, INTEGER const 
         //
         lstres = zero;
         for (i = 1; i <= n; i = i + 1) {
-            lstres = max(lstres, REAL(abs(x[(i - 1) + (j - 1) * ldx])));
+            lstres = max(lstres, abs(x[(i - 1) + (j - 1) * ldx]));
         }
         if (lstres != zero) {
             ferr[j - 1] = ferr[j - 1] / lstres;
