+++ mplapack/reference/Rlasyf.cpp
@@ -38,8 +38,8 @@
 
 void Rlasyf(const char *uplo, INTEGER const n, INTEGER const nb, INTEGER &kb, REAL *a, INTEGER const lda, INTEGER *ipiv, REAL *w, INTEGER const ldw, INTEGER &info) {
     const REAL one = 1.0;
-    const REAL sevten = 17.0e+0;
-    const REAL eight = 8.0e+0;
+    const REAL sevten = 17.0;
+    const REAL eight = 8.0;
     REAL alpha = 0.0;
     INTEGER k = 0;
     INTEGER kw = 0;
@@ -145,7 +145,7 @@ void Rlasyf(const char *uplo, INTEGER const n, INTEGER const nb, INTEGER &kb, RE
                 rowmax = abs(w[(jmax - 1) + ((kw - 1) - 1) * ldw]);
                 if (imax > 1) {
                     jmax = iRamax(imax - 1, &w[((kw - 1) - 1) * ldw], 1);
-                    rowmax = max(rowmax, REAL(abs(w[(jmax - 1) + ((kw - 1) - 1) * ldw])));
+                    rowmax = max(rowmax, abs(w[(jmax - 1) + ((kw - 1) - 1) * ldw]));
                 }
                 //
                 if (absakk >= alpha * colmax * (colmax / rowmax)) {
@@ -435,7 +435,7 @@ void Rlasyf(const char *uplo, INTEGER const n, INTEGER const nb, INTEGER &kb, RE
                 rowmax = abs(w[(jmax - 1) + ((k + 1) - 1) * ldw]);
                 if (imax < n) {
                     jmax = imax + iRamax(n - imax, &w[((imax + 1) - 1) + ((k + 1) - 1) * ldw], 1);
-                    rowmax = max(rowmax, REAL(abs(w[(jmax - 1) + ((k + 1) - 1) * ldw])));
+                    rowmax = max(rowmax, abs(w[(jmax - 1) + ((k + 1) - 1) * ldw]));
                 }
                 //
                 if (absakk >= alpha * colmax * (colmax / rowmax)) {
