+++ mplapack/reference/Rsytf2.cpp
@@ -39,8 +39,8 @@
 void Rsytf2(const char *uplo, INTEGER const n, REAL *a, INTEGER const lda, INTEGER *ipiv, INTEGER &info) {
     bool upper = false;
     const REAL one = 1.0;
-    const REAL sevten = 17.0e+0;
-    const REAL eight = 8.0e+0;
+    const REAL sevten = 17.0;
+    const REAL eight = 8.0;
     REAL alpha = 0.0;
     INTEGER k = 0;
     INTEGER kstep = 0;
@@ -141,7 +141,7 @@ void Rsytf2(const char *uplo, INTEGER const n, REAL *a, INTEGER const lda, INTEG
                 rowmax = abs(a[(imax - 1) + (jmax - 1) * lda]);
                 if (imax > 1) {
                     jmax = iRamax(imax - 1, &a[(imax - 1) * lda], 1);
-                    rowmax = max(rowmax, REAL(abs(a[(jmax - 1) + (imax - 1) * lda])));
+                    rowmax = max(rowmax, abs(a[(jmax - 1) + (imax - 1) * lda]));
                 }
                 //
                 if (absakk >= alpha * colmax * (colmax / rowmax)) {
@@ -311,7 +311,7 @@ void Rsytf2(const char *uplo, INTEGER const n, REAL *a, INTEGER const lda, INTEG
                 rowmax = abs(a[(imax - 1) + (jmax - 1) * lda]);
                 if (imax < n) {
                     jmax = imax + iRamax(n - imax, &a[((imax + 1) - 1) + (imax - 1) * lda], 1);
-                    rowmax = max(rowmax, REAL(abs(a[(jmax - 1) + (imax - 1) * lda])));
+                    rowmax = max(rowmax, abs(a[(jmax - 1) + (imax - 1) * lda]));
                 }
                 //
                 if (absakk >= alpha * colmax * (colmax / rowmax)) {
